requirements.txt
# Core async runtime
aiohttp>=3.9.0
websockets>=12.0
asyncio-throttle>=1.0.2

# Exchange SDKs
ccxt>=4.3.0                     # Toobit + 100+ other exchanges (async support)
hyperliquid-python-sdk>=0.9.0
lighter-sdk>=1.0.8
# py-clob-client for Polymarket (install manually if Polymarket trading desired)

# Technical analysis
pandas>=2.1.0
numpy>=1.26.0
ta>=0.11.0
pandas-ta>=0.3.14b

# Crypto / blockchain
eth-account>=0.10.0
web3>=6.15.0

# Config & persistence
pyyaml>=6.0.1
python-dotenv>=1.0.0
aiosqlite>=0.19.0

# Monitoring & alerting
python-telegram-bot>=20.7
rich>=13.7.0
colorlog>=6.8.2

# Web dashboard (gamified browser UI)
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Claude AI trade advisor (optional — enables AI signal review)
anthropic>=0.25.0

# Utilities
python-dateutil>=2.8.2
humanize>=4.9.0
orjson>=3.9.10
httpx>=0.26.0
tenacity>=8.2.3
